% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_both.R
\name{jp_both}
\alias{jp_both}
\alias{print.jp_both}
\title{Amazing Random Selection About Japanese Given Names}
\usage{
jp_both(ind)

\method{print}{jp_both}(x, width = NULL, ...)
}
\arguments{
\item{x}{Default object for \code{print} method}

\item{width}{Optional column width parameter}

\item{...}{Other optional arguments}
}
\description{
Amazing Random Selection About Japanese Given Names
}
\details{
Extract textual information
}

