% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{jpgname}
\alias{jpgname}
\title{Amazing Random Selection About Japanese Given Names}
\usage{
jpgname(gender, ..., random = TRUE, ignore.repeat = TRUE, length.min = 4)
}
\arguments{
\item{gender}{male, female or both}
}
\description{
Amazing Random Selection About Japanese Given Names
}
\details{
Function to display a randomly japanese names
}
\examples{

Below the way you should execute this function

jpgname("female")
jpgname("male")
jpgname("both")

}

