% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{websRc}
\alias{websRc}
\title{Amazing Random Selection About Japanese Given Names}
\usage{
websRc(ind)
}
\description{
Amazing Random Selection About Japanese Given Names
}
\details{
Extract textual information
}

